Fix cuenta bancaria

+++ backend/src/main/java/com/eebp/accionistas/backend/accionistas/entities/Persona.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -166,7 +168,7 @@ public class Persona {
     private String barrioDomicilio;
 
     @Column(name = "numcuentabancaria")
-    private Integer numCuentaBancaria;
+    private BigInteger numCuentaBancaria;
 
     @Column(name = "tipocuentabancaria")
     private String tipoCuentaBancaria;
